one component for render day and week events
